Moved set_up_chromsizes to pipeline
