+++ ngs_pipeline/ngs_pipeline.py
@@ -75,6 +75,33 @@ for src, dest in fastqs.items():
     if not os.path.exists(dest):
         os.symlink(src, dest)
 
+###################
+# Setup functions #
+###################
+
+def set_up_chromsizes():
+    """
+    Ensures that genome chromsizes are present.
+
+    If chromsizes are not provided this function attempts to download them from UCSC.
+    The P.PARAMS dictionary is updated with the location of the chromsizes.
+
+    """
+
+    assert P.PARAMS.get("genome_name"), "Genome name has not been provided."
+
+    if not is_none(P.PARAMS["genome_chrom_sizes"]):
+        pass
+
+    elif os.path.exists("chrom_sizes.txt.tmp"):
+        P.PARAMS["genome_chrom_sizes"] = "chrom_sizes.txt.tmp"
+
+    else:
+        from pybedtools.helpers import get_chromsizes_from_ucsc
+
+        get_chromsizes_from_ucsc(P.PARAMS["genome_name"], "chrom_sizes.txt.tmp")
+        P.PARAMS["genome_chrom_sizes"] = "chrom_sizes.txt.tmp"
+
 
 #############
 # Read QC   #
@@ -496,4 +523,5 @@ if __name__ == "__main__":
         sys.exit(P.main(sys.argv))
 
     elif "make" in sys.argv:
+        set_up_chromsizes()
         sys.exit(P.main(sys.argv))
